massive amount of bug fixing and feature addition, now a preliminary version of BScope compiles and acquires 2D and 3D data.

+++ RTlib-kam/src/rtlib/kam/memory/cursor/NDCursor.java
@@ -2,7 +2,7 @@ package rtlib.kam.memory.cursor;
 
 public interface NDCursor
 {
-	public long getDimension();
+	public int getDimension();
 
 	public long[] getDimensions();
